requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
numpy==2.0.1
    # via opencv-python
opencv-python==4.10.0.84
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
